Fix scrolling on multi-monitor setups
WM_MOUSEWHEEL gives mouse coordinates in screen space, as signed 16-bit
numbers, but my code was treating the numbers as unsigned 16-bit
numbers. This caused the right split to scroll when trying to scroll the
left view when the am7 window was on a monitor to the left of the main monitor,
because that monitor would have negative screen coordinates.