Fix occasional crash when copying multiple selections
This was happening because I hadn't realized that buffer_get_slice will
use the stack to return a contiguous array if we happen to request
something which straddles the buffer gap.

But we also depended on the exclusively using the stack for
concatenating the returned strings to build 'full'.

So, if we have e.g. the following code:
    x
    y
And set up a cursor at the start of both lines and type to get the
following:
    foo.x
    foo.y
And then we tried to select the entirety of both lines and copy, we
would crash. This is because the gap would be between "foo." and "y", so
the second call to buffer_get_slice would use the stack to allocate the
return value, which would mess up 'full' and cause 'stack_printf_append'
to fail.